scap: Enable require_valid_service
This way, there won't be a paradox when provisioning a new machine,
whereby Puppet clones the source code, and initialises scap, which
then tries to restart the service that doesn't exist yet, because
Puppet won't create the service until after Package[] resource
is fulfilled, which includes the scap-init step.

Enabling require_valid_service breaks the cycle by having the
scap-init step gracefully skip restarting of the service when
it doesn't exist yet. Thus, allowing Puppet to create the service
and start it (the first time), after which on subsequent deploys
scap will be able to restar the service.

Bug: T195314
Change-Id: I71ecc6db804ca3a5a8251bbb8f0f3b70f06c99f8