navtiming: Log deprecation message when seeing legacy userAgent fields
Also remove the test cases, given we won't change this code anymore,
the next commit removes it.

This is so that the tests may complete without a sea of deprecation
warnings, and to make it easy to confirm that the remaining code
indeed does not wrongly trigger this during tests.

Bug: T212537
Change-Id: I60350563c0ad3602d1f48c18c9ad4625c591e379