Remove “site” field from FID and PaintTiming
It turns out to be overridden in production Prometheus data
once consumed in Grafana (replaced by the name of the DC
collecting the Prometheus metrics).

Furthermore, it’s of little use since we can already differentiate
mobile and desktop browsers. The amount of mobile users using
the desktop site and vice versa is negligible.

Bug: T175087
Change-Id: I22c69b64ac1e0a7b0b6774d9ee564f595f5a4170