Fix tests that just passed on Python 3.
This makes sure our expect tests actually runs in Python 3 instead of
just passing independently.

This is what I think has happened: When we upgraded to Python 3, the
unit test function assertItemsEqual was removed, and we fixed that
with a try/catch and pass all tests that was catched.
See https://bugs.python.org/issue17866

However assertItemsEqual become assertCountEqual but it didn't work
out of the box since Python 3 changed so when our key string become a
byte string and couldn't be compared with our other string.

Bug: T323275
Change-Id: I16a9da1710f30359b25da0e05f904851b445cc6a