Switch remainig use of `frontend.*_discard` to `eventlogging.client_errors`
* eventlogging.client_errors

Used on two dashboards:
1. <https://grafana.wikimedia.org/d/000000494/eventlogging-schema-jumbo?var-schema=NavigationTiming>
2. <https://grafana.wikimedia.org/d/000000143/navigation-timing?viewPanel=22>

This is the main place where we want to record invalid/rejected events
as it doesn't require knowledge of any navtiming-specific metric
namespaces and thus gets displayed on the general EventLogging dashboard
when you select our schema. We also display the same data on the NT dashboard
(link 2 above).

* frontend.painttiming_discard

Used nowhere. Fixed in this patch by sending the data to the standard
client_errors prefix instead, so that it will start showing up where
we want it to show up.

> $ node search-grafana-dashboards.js 'painttiming_discard'
> Checked 704 dashboards. Found 0 results.

* frontend.navtiming_discard

Used on 1 dashboard:
* <https://grafana.wikimedia.org/d/000000143/navigation-timing?viewPanel=22>

It is queried htere for compatibility with old data in the same panel
that also queries "eventlogging.client_errors". We don't need to send
it to both forever though, so this patch switches the remaining "isSane"
error to the new prefix as well.

Bug: T286703
Change-Id: Iba30c585ed2c742da1c1d677ef69f2b3a3e26205