Apply new buckets to navtiming_responsestart_by_cache_host also
This one didn't yet have buckets set, so it used the default which has
15 buckets instead of the 10 buckets we choose for the (new) Prometheus
metrics that represent early events with relatively low numbers.

The main difference is that our buckets start at 10ms instead of 5ms,
and use +Inf after 2.5s instead of after 10s which seems fine for
responseStart. Current Prometeheus data for this metric shows <0.1%
in both the le=5ms and le=10ms buckets, with 0.99 quantile around
3s.

Also:

* Rename from early/late to low/high to make it reflect the range
  of values instead of where it is in the page load cycle.

* Rename navtiming_ttfb to responsestart for consistency with
  navtiming_responsestart_by_cache_host and other navtiming metrics.

* Switch most deltas from high to low. onLoad, while late, is a very
  small delta from loadEventStart to loadEventEnd.

* Tentatively remove the commented-out 'request' and 'response' deltas.

Bug: T326118
Change-Id: I987e8376de6c50d6ba72896e67e61c1650c4a9f0