Remove user agent and battery and add platform for CPU benchmark.
I don't see the use case of battery, if the performance differs
we will do anyhting about it for now. The same with the user agent.
We still collect that data and can dig into. Instead we should use
the Prometheus so we can compare the becnhmark between countries
and what I really after is to use it to calibrate our synthetic
tests. That's why I also added platform so we a little better way
to know if the benchmark is on desktop or mobile.

Bug: T321398
Change-Id: Ia7f1caa517b87484667ece0c1b513d0667c5e23a